feat: update posts by ID
