+++ src/controllers/posts.controller.js
@@ -5,20 +5,24 @@ export const createPost = async (req, res) => {
     const { hashtags } = res.locals;
 
     try {
-        await postsService.createPost(req.body, hashtags, /* userID */);
+        await postsService.createPost(req.body, hashtags, 1);
         res.sendStatus(201);
 
-    } catch (err){
+    } catch (err) {
         res.send(err.message);
     }
 }
 
 export const updatePost = async (req, res) => {
 
+    const { hashtags } = res.locals;
+    const { postID } = req.params;
+
     try {
+        await postsService.updatePost(req.body, hashtags, postID, 1);
+        res.sendStatus(200);
 
-
-    } catch (err){
+    } catch (err) {
         res.send(err.message);
     }
 }
@@ -26,16 +30,12 @@ export const updatePost = async (req, res) => {
 export const deletePost = async (req, res) => {
 
     const { postID } = req.params;
-    if (!Number(postID)) {
-        return res.sendStatus(403);
-    }
 
     try {
-
-        await postsService.deletePost(Number(postID), /* userID */);
+        await postsService.deletePost(postID, 1);
         res.sendStatus(204);
 
-    } catch (err){
+    } catch (err) {
         res.send(err.message);
     }
 }
