+++ src/routes/posts.routes.js
@@ -8,13 +8,13 @@ import {
 
 import { schemaValidation } from "../middlewares/schemaValidation.js";
 import { getHashtags } from "../middlewares/getHashtags.js";
-
+import { checkPostID } from "../middlewares/checkPostID.js";
 import { postsSchema } from "../schemas/posts.schema.js";
 
 const postsRouter = Router();
 
 postsRouter.post("/create-post", schemaValidation(postsSchema), getHashtags, createPost);
-postsRouter.put("/update-post/:postID", schemaValidation(postsSchema), getHashtags, updatePost);
-postsRouter.delete("/delete-post/:postID", deletePost);
+postsRouter.put("/update-post/:postID", checkPostID, schemaValidation(postsSchema), getHashtags, updatePost);
+postsRouter.delete("/delete-post/:postID", checkPostID, deletePost);
 
 export default postsRouter;
