+++ src/services/posts.service.js
@@ -1,5 +1,16 @@
 import db from "../database/db.connection.js";
-import hashtagService from "./hashtag.service.js";
+import hashtagsService from "./hashtags.service.js";
+
+const checkUserPost = async (postID, userID) => {
+
+    const result = await db.query(
+        `SELECT * FROM posts
+         WHERE id = $1 AND "userID" = $2;
+        `, [postID, userID]
+    );
+
+    return result;
+}
 
 const createPost = async (payload, hashtags, userID) => {
 
@@ -19,19 +30,8 @@ const createPost = async (payload, hashtags, userID) => {
 
     if (hashtags.length > 0) {
         const postID = createdPost.rows[0].id;
-        hashtagService.createAndInsertHashtags(hashtags, postID);
+        hashtagsService.createAndInsertHashtags(hashtags, postID);
     }
-}
-
-const checkUserPost = async (postID, userID) => {
-
-    const result = await db.query(
-        `SELECT * FROM posts
-         WHERE id = $1 AND "userID" = $2
-        `, [postID, userID]
-    );
-
-    return result;
 }
 
 const deletePost = async (postID, userID) => {
@@ -41,19 +41,55 @@ const deletePost = async (postID, userID) => {
         return;
     }
 
-    hashtagService.deleteHashtags(postID);
+    const resultHashtagPost = await hashtagsService.getHashtagsByPost(postID);
+    if (resultHashtagPost.rows.length > 0) {
+        hashtagsService.deleteHashtags(resultHashtagPost, postID);
+    }
 
     await db.query(
         `DELETE FROM posts
-         WHERE id = $1 AND "userID" = $2
+         WHERE id = $1 AND "userID" = $2;
         `, [postID, userID]
-    )
+    );
+}
+
+const updatePost = async (payload, hashtags, postID, userID) => {
+    
+    const result = await checkUserPost(postID, userID);
+    if (!result.rows[0]) {
+        return;
+    }
+
+    const resultHashtagPost = await hashtagsService.getHashtagsByPost(postID);
+    if (resultHashtagPost.rows.length > 0) {
+        hashtagsService.deleteHashtags(resultHashtagPost, postID);
+    }
+
+    const {
+        description,
+        URL,
+    } = payload;
+
+    await db.query(
+        `UPDATE posts SET
+            description = $1,
+            "URL" = $2,
+            "userID" = $3
+         WHERE id = $4
+        `, [description, URL, userID, postID]
+    );
+
+    if (hashtags.length > 0) {
+        hashtagsService.createAndInsertHashtags(hashtags, postID);
+    }
 }
 
 
 const postsService = {
+    checkUserPost,
     createPost,
-    deletePost
+    deletePost,
+    updatePost
 }
 
 export default postsService;
